INTDEV-733: Delete Template and CardContainer
Template's 609 lines were card-tree operations plus path resolution that
TemplateResource already performs, reached through a `templateObject()`
indirection that existed only to hold them. They move onto
TemplateResource, which owns the template's name, its folder and its
schema anyway:

- cards()/listCards() -> templateCards(); attachments() ->
  templateAttachments(); findCard() -> templateCard(); addCard/addCards
  and createCards keep their names. The unused placeholder path
  parameter on cards() goes with the base class it was there for.
- cardFolder and cardAttachmentFolder do not come along:
  project.cardFolder(key) and project.cardAttachmentFolder(key) route by
  key to the tree that holds the card and answer the same question, for
  a template card as for a project one.
- setTemplatePath and moduleTemplatePath are gone outright, with their
  'optimization - if template.path is set - use it' branch and the
  module-folder scan behind it: the resource's own internalFolder is the
  answer they were reconstructing, including for module templates.
- findCard's prefix-mismatch guard (a module card asked of a local
  template, and the reverse) goes with them, and so does the linear scan
  over cards() it fronted. The tree lookup is exact, so the impossible
  answer the guard screened for cannot occur; what a card outside the
  template gets back is the same error as before.
- rootLevelProjectCards, which hydrated every project card to filter it
  down to the root ones, becomes project.showProjectCards().

CardContainer's remaining members were Project's: the file-name statics,
the card read/write helpers, and hasCard. They move into Project, which
no longer extends anything, and the helpers no protected member needs
become private. hasProjectCard and hasTemplateCard do not come along -
the three call sites that asked "is this a template card?" were really
asking "does this card take part in a workflow", which is the tree's
validationApplies flag, so they now ask that instead. That gives the flag
its second consumer (permission checks, alongside metadata validation)
and removes the last location discriminator from the command layer.

Also gone: Project.templateObjectByName, whose one caller (Move) asks
templateResource for the same resource, and
Project.createTemplateObjectFromCard, which only a test called - the
tree that holds a card names its container, so nothing needs a container
object built from one.

Behavior change vs main: the template group in listCards() is named by
the resource's fullName instead of by its data?.name. That is the same
string for a template whose file has been read, and no longer the empty
string for one whose file has not.

Test changes, all of them mechanical accessor renames with no assertion
changed unless said otherwise:
- test/template.test.ts and test/template-create-cards.test.ts
  constructed Template directly; they now construct the TemplateResource
  (which is what the deleted class delegated to) and call the renamed
  methods. Both files' assertions are unchanged, including the three
  bug-fix regressions in template-create-cards.test.ts (template cards
  unchanged after a double instantiation; the original error surviving a
  failed write with no partial cards; cached content and attachments
  matching what was written). The two path assertions that went through
  the deleted cardFolder/cardAttachmentFolder now go through the
  project's key-routed reads, and 'check all the attachments' asserts
  that the project's answer is the template tree's own
  attachmentFolderOf.
- test/card-tree.test.ts, test/command-create.test.ts,
  test/command-remove.test.ts, test/project.test.ts:
  '.templateObject()' dropped from the call chain, '.cards()' ->
  '.templateCards()'. project.test.ts's hasTemplateCard assertion becomes
  the tree identity it stood for: treeOf(key) is the project's own tree,
  and its createTemplateObjectFromCard assertion becomes
  templateResource(treeOf(key).name).

No test added, deleted or weakened: 122 files | 1839 passed | 1 skipped,
the same as the previous commit.
