+++ tools/data-handler/src/commands/create.ts
@@ -115,14 +115,13 @@ export class Create {
       templateName,
       'templates',
     );
-    const templateObject = templateResource.templateObject();
-    const specificCard = card ? templateObject.findCard(card) : undefined;
+    const specificCard = card ? templateResource.templateCard(card) : undefined;
 
-    if (isModulePath(templateObject.templateFolder())) {
+    if (isModulePath(templateResource.templateFolder())) {
       throw new Error(`Cannot add cards to imported module templates`);
     }
 
-    const cardsContainer = await templateObject.addCards(
+    const cardsContainer = await templateResource.addCards(
       cardTypeName,
       count,
       specificCard,
@@ -207,12 +206,11 @@ export class Create {
     const specificCard = parentCardKey
       ? this.project.findCard(parentCardKey)
       : undefined;
-    const templateObject = templateResource.templateObject();
-    if (!templateObject || !templateObject.isCreated()) {
+    if (!templateResource.isCreated()) {
       throw new Error(`Template '${templateName}' not found from project`);
     }
 
-    const createdCards = await templateObject.createCards(specificCard);
+    const createdCards = await templateResource.createCards(specificCard);
     if (createdCards.length > 0) {
       await this.project.runCreationSideEffects(
         createdCards.map((card) => card.key),
