+++ tools/data-handler/src/commands/import.ts
@@ -198,12 +198,7 @@ export class Import {
         template,
         'templates',
       );
-      const templateObject = templateResource.templateObject();
-      if (!templateObject) {
-        throw new Error(`Template '${template}' not found`);
-      }
-
-      const templateCards = templateObject.cards();
+      const templateCards = templateResource.templateCards();
       if (templateCards.length !== 1) {
         console.warn(
           `Template '${template}' for card '${title}' does not have exactly one card. Skipping row.`,
