+++ tools/data-handler/src/mutations/handlers/card-type-rename.ts
@@ -107,7 +107,7 @@ export class CardTypeRenameHandler implements Handler<RenameInput> {
     const project = [...ctx.project.cards(undefined)];
     const templates = ctx.project.resources
       .templates(ResourcesFrom.localOnly)
-      .flatMap((t) => t.templateObject().cards());
+      .flatMap((t) => t.templateCards());
     return [...project, ...templates].filter(
       (c) => c.metadata?.cardType === oldName,
     );
