+++ tools/data-handler/src/mutations/handlers/card-type-workflow-change.ts
@@ -70,7 +70,7 @@ export class CardTypeWorkflowChangeHandler implements Handler<EditInput> {
     const project = [...ctx.project.cards(undefined)];
     const templates = ctx.project.resources
       .templates(ResourcesFrom.localOnly)
-      .flatMap((t) => t.templateObject().cards());
+      .flatMap((t) => t.templateCards());
     return [...project, ...templates].filter(
       (c) => c.metadata?.cardType === cardTypeName,
     );
