+++ tools/data-handler/src/mutations/handlers/field-type-enum-rename.ts
@@ -61,7 +61,7 @@ export class FieldTypeEnumRenameHandler implements Handler<EditInput> {
     const project = [...ctx.project.cards(undefined)];
     const templates = ctx.project.resources
       .templates(ResourcesFrom.localOnly)
-      .flatMap((t) => t.templateObject().cards());
+      .flatMap((t) => t.templateCards());
     return [...project, ...templates].filter(
       (c) => c.metadata && fieldName in c.metadata,
     );
