+++ tools/data-handler/test/template.test.ts
@@ -10,7 +10,6 @@ import { copyDir } from '../src/utils/file-utils.js';
 import { getTestProject } from './helpers/test-utils.js';
 import type { Project } from '../src/containers/project.js';
 import { resourceName } from '../src/utils/resource-utils.js';
-import { Template } from '../src/containers/template.js';
 import { TemplateResource } from '../src/resources/template-resource.js';
 import { CardNotFoundError } from '../src/exceptions/index.js';
 
@@ -32,50 +31,43 @@ afterAll(() => {
   rmSync(testDir, { recursive: true, force: true });
 });
 
+// Constructed directly: resources.byType only knows templates that are on disk.
+function templateOf(name: string): TemplateResource {
+  return new TemplateResource(project, resourceName(name));
+}
+
 describe('template', () => {
   it('try to create template with no name', () => {
-    expect(() => new Template(project, { name: '', path: '' })).toThrow(
+    expect(() => templateOf('')).toThrow(
       `Must define resource name to query its details`,
     );
   });
   it('show template cards', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/simplepage',
-      path: '',
-    });
-    const cards = template.cards();
+    const template = templateOf('decision/templates/simplepage');
+    const cards = template.templateCards();
     expect(cards.length).toBe(3);
   });
   it('show template cards from empty template', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/empty',
-      path: '',
-    });
-    const cards = template.cards();
+    const template = templateOf('decision/templates/empty');
+    const cards = template.templateCards();
     expect(cards.length).toBe(0);
   });
 
   it('throws an error when creating cards from an empty template', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/empty',
-      path: '',
-    });
+    const template = templateOf('decision/templates/empty');
     await expect(template.createCards()).rejects.toThrow(Error);
-    expect(template.cards().length).toBe(0);
+    expect(template.templateCards().length).toBe(0);
   });
   it('create template card under a specific card from a project', async () => {
     // Choose specific card so that it does not have currently child cards.
     const cardBefore = project.findCard('decision_6');
     expect(cardBefore?.children?.length).toBe(0);
 
-    const template = new Template(project, {
-      name: 'decision/templates/simplepage',
-      path: '',
-    });
+    const template = templateOf('decision/templates/simplepage');
 
     // Check that created cards are mapped from template cards.
     const createdCards = await template.createCards(cardBefore);
-    const templateCards = template.cards();
+    const templateCards = template.templateCards();
 
     expect(
       createdCards.map((item) => item.metadata!.templateCardKey),
@@ -86,10 +78,7 @@ describe('template', () => {
     expect(cardAfter?.children?.length).toBe(2);
   });
   it('throws an error when trying to create a specific card from an empty template', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/empty',
-      path: '',
-    });
+    const template = templateOf('decision/templates/empty');
     const nonExistingCard: Card = {
       key: '1111',
       path: '',
@@ -99,14 +88,11 @@ describe('template', () => {
     };
 
     await expect(template.createCards(nonExistingCard)).rejects.toThrow(Error);
-    expect(template.cards().length).toBe(0);
+    expect(template.templateCards().length).toBe(0);
   });
 
   it('creates no cards when the template is not empty and parent is non-existent', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/simplepage',
-      path: '',
-    });
+    const template = templateOf('decision/templates/simplepage');
     const nonExistingCard: Card = {
       key: '1111',
       path: '',
@@ -124,45 +110,30 @@ describe('template', () => {
   });
 
   it('add new card to a template', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
-    const cardsBefore = template.cards();
+    const template = templateOf('decision/templates/decision');
+    const cardsBefore = template.templateCards();
     await template.addCard('decision/cardTypes/decision');
-    const cardsAfter = template.cards();
+    const cardsAfter = template.templateCards();
     expect(cardsBefore.length + 1).toBe(cardsAfter.length);
   });
   it('list attachments from a template (no attachments in template)', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/simplepage',
-      path: '',
-    });
-    const attachments = template.attachments();
+    const template = templateOf('decision/templates/simplepage');
+    const attachments = template.templateAttachments();
     expect(attachments.length).toBe(0);
   });
   it('list attachments from a template', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
-    const attachments = template.attachments();
+    const template = templateOf('decision/templates/decision');
+    const attachments = template.templateAttachments();
     expect(attachments.length).toBe(1);
   });
   it('list attachments from an empty template', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/empty',
-      path: '',
-    });
-    const attachments = template.attachments();
+    const template = templateOf('decision/templates/empty');
+    const attachments = template.templateAttachments();
     expect(attachments.length).toBe(0);
   });
   it('check that template does not exist, then create it', async () => {
     const templateName = 'decision/templates/idontexistyet';
-    const template = new Template(project, {
-      name: templateName,
-      path: '',
-    });
+    const template = templateOf(templateName);
 
     expect(template.isCreated()).toBe(false);
 
@@ -175,23 +146,17 @@ describe('template', () => {
     expect(template.isCreated()).toBe(true);
   });
   it('check template paths', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
     const templateMain = template.templateFolder();
     const templateCards = template.templateCardsFolder();
-    const specificCardPath = template.cardFolder('decision_1');
+    const specificCardPath = await project.cardFolder('decision_1');
     expect(templateMain).toContain('.cards');
     expect(join(templateMain, 'c')).toBe(templateCards);
     expect(templateCards).toContain(`decision${sep}c`);
     expect(specificCardPath).toContain(`decision${sep}c${sep}decision_1`);
   });
   it('add card to a template', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
     const parentCard: Card = {
       key: 'decision_1',
       path: join(template.templateCardsFolder(), 'decision_1'),
@@ -203,20 +168,17 @@ describe('template', () => {
     ).resolves.not.toThrow();
   });
   it('access card details by id', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
     const cardToOperateOn = 'decision_1';
     const cardExists = template.hasTemplateCard(cardToOperateOn);
     expect(cardExists).toBe(true);
 
-    const card = template.findCard(cardToOperateOn);
+    const card = template.templateCard(cardToOperateOn);
     expect(card).not.toBeUndefined();
     expect(card.metadata!.title).toBe('Untitled');
     expect(card.metadata!.cardType).toBe('decision/cardTypes/decision');
     expect(card.metadata!.workflowState).toBe('Draft');
-    const additionalCardDetails = template.findCard(cardToOperateOn);
+    const additionalCardDetails = template.templateCard(cardToOperateOn);
     expect(additionalCardDetails).not.toBeUndefined();
     expect(additionalCardDetails.metadata!.title).toBe('Untitled');
     expect(additionalCardDetails.metadata!.cardType).toBe(
@@ -228,28 +190,19 @@ describe('template', () => {
     expect(additionalCardDetails.content).not.toBeUndefined();
   });
   it('try to add card to a template that does not exist on disk', async () => {
-    const template = new Template(project, {
-      name: 'i-dont-exist',
-      path: '',
-    });
+    const template = templateOf('i-dont-exist');
 
     await expect(
       template.addCard('decision/cardTypes/decision'),
     ).rejects.toThrow();
   });
   it('try to add card to a template from card type that does not exist', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
 
     await expect(template.addCard('i-dont-exist')).rejects.toThrow();
   });
   it('try to add card to a template to a parent card that does not exist', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
     const parentCard: Card = {
       key: 'i-dont-exist',
       path: join(template.templateCardsFolder(), 'decision_1'),
@@ -262,25 +215,23 @@ describe('template', () => {
     ).rejects.toThrow();
   });
   it('check all the attachments', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
 
     // Project can fetch the template's attachment's folder.
-    const attachmentFolder1 = project.cardAttachmentFolder('decision_1');
-    const attachmentFolder2 = template.cardAttachmentFolder('decision_1');
-    expect(attachmentFolder1).toContain('decision_1');
-    expect(attachmentFolder1).toContain(sep + 'a');
-    expect(attachmentFolder1).toBe(attachmentFolder2);
+    const attachmentFolder = project.cardAttachmentFolder('decision_1');
+    expect(attachmentFolder).toContain('decision_1');
+    expect(attachmentFolder).toContain(sep + 'a');
+    expect(attachmentFolder).toBe(
+      template.cardTree.attachmentFolderOf('decision_1'),
+    );
 
-    expect(() => template.cardAttachmentFolder('decision_999')).toThrow(
+    expect(() => project.cardAttachmentFolder('decision_999')).toThrow(
       CardNotFoundError,
     );
 
-    const templateAttachments = template.attachments();
+    const templateAttachments = template.templateAttachments();
     expect(templateAttachments.length).toBe(1);
-    const templateCard = template.findCard('decision_1');
+    const templateCard = template.templateCard('decision_1');
     const cardAttachments = templateCard.attachments;
     expect(cardAttachments.at(0)!.card).toBe('decision_1');
     expect(cardAttachments.at(0)!.fileName).toBe('the-needle.heic');
@@ -288,36 +239,24 @@ describe('template', () => {
     expect(cardAttachments.at(0)!.path).toContain(sep + 'a');
   });
   it('check if template is created', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
-    const nonExistingTemplate = new Template(project, {
-      name: 'idontExist',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
+    const nonExistingTemplate = templateOf('idontExist');
 
     expect(template.isCreated()).toBe(true);
     expect(nonExistingTemplate.isCreated()).toBe(false);
   });
   it('find certain card from template', () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
+    const template = templateOf('decision/templates/decision');
 
     expect(() => {
-      template.findCard('idontexist');
+      template.templateCard('idontexist');
     }).toThrow(`Card 'idontexist' is not part of template`);
 
-    const existingCard = template.findCard('decision_1');
+    const existingCard = template.templateCard('decision_1');
     expect(existingCard).not.toBeUndefined();
   });
   it('show template details', async () => {
-    const template = new TemplateResource(
-      project,
-      resourceName('decision/templates/decision'),
-    );
+    const template = templateOf('decision/templates/decision');
 
     const templateDetails = template.show();
     expect(templateDetails.name).toBe('decision/templates/decision');
@@ -328,11 +267,8 @@ describe('template', () => {
     expect(templateDetails.displayName).toBe('Decision');
   });
   it('list template cards with card keys', async () => {
-    const template = new Template(project, {
-      name: 'decision/templates/decision',
-      path: '',
-    });
-    const templateCards = template.listCards();
+    const template = templateOf('decision/templates/decision');
+    const templateCards = template.templateCards();
     expect(templateCards.length).toBeGreaterThan(0);
   });
 });
